Fixed the menu selection that was letting garbage
This issue was introduced by the fix for another menu issue that was crashing
Wings3D when closing a window with the context menu active (commit: #be68730)

The fix redraw the entire menu unselected before it be displayed. It was also
needed to find for the window of the submenu item because it's not stored in
the object property of menudata.

NOTE: Fixed the menu selection that was letting garbage. Thanks to sciroccorics