Updated the Align to Target command adding new options
It was discussed about this command at Discord where some users asked for
it to be used to directly align an object to the regular axis (X,Y,Z) working
similar to PutOn. I realise when I first implemented it that it was a missing
feature that could speeded up that operation.

I also reorgnized the code by better naming the functions and moving the
option to the context menu for body selection mode - just next to the
Absolute Commands, instead of inside it. That give it more visibility to new
users and also seems to be the appropriated place by considering the kind of
operations that are under that option.

NOTE: Updated the 'Align To Target' command with new options. It's now called
'Align To' and it was put out of the 'Absolute Commands' option. Thanks to
envelupo, Vaughan, Arclite and LLugnuto779